+++ views/components/certification/CertIcon.js
