+++ views/components/certification/Item.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components/macro'
+
+const Item = styled.li`
+  position: absolute;
+  top: 0;
+  display: block;
+  width: 25%;
+  height: 100%;
+  background-color: rgba(255, 255, 255, 0);
+  background-repeat: repeat;
+  margin-left: -12.5%;
+`
+
+export default Item
